pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "panoptes_aggregation"
description = "Aggregation code for Zooniverse panoptes projects."
authors = [
    {name = "Coleman Krawczyk", email = "coleman@zooniverse.org"},
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: Apache Software License"
]
dynamic = ["version"]
requires-python = ">=3.8,<3.12"
dependencies = [
    "beautifulsoup4>=4.8.1,<4.13",
    "collatex@git+https://github.com/interedition/collatex.git@174035dda549ac3f967919a8475fdeb8dbb48f40#subdirectory=collatex-pythonport",
    "hdbscan>=0.8.20,<=0.8.33",
    "lxml>=4.4,<4.10",
    "numpy>=1.22.0,<1.25.3",
    "packaging>=20.1,<23.2",
    "pandas>=1.4.0,<2.1.1",
    "progressbar2>=3.39,<4.3",
    "python-levenshtein>=0.21.0,<0.22",
    "python-slugify>=7.0.0,<8.1",
    "pyyaml>=6.0,<6.1",
    "scikit-learn>=1.2.0,<1.3.1",
    "scipy>=1.10.0,<1.10.2",
    "werkzeug>=2.3.0,<2.3.8",
    "shapely>=2.0,<2.0.2"
]

[project.optional-dependencies]
online = [
    "flask>=2.0,<2.4",
    "flask-cors>=3.0,<3.1",
    "panoptes-client>=1.6,<1.7",
    "requests>=2.28,<2.32",
    "gunicorn>=20.0,<20.2",
    "sentry-sdk[flask]>=1.0,<1.30",
    "newrelic>=8.4.0,<8.8.1",
    "gitpython>=3.0.0,<3.2"
]
gui = [
    "Gooey>=1.0.8.1,<1.1"
]
doc = [
    "matplotlib>=3.5.1,<3.8",
    "myst-nb>=0.13.2,<0.18",
    "sphinx>=5.2.0,<7.3",
    "sphinxcontrib-httpdomain>=1.7.0,<1.9",
    "sphinx_rtd_theme>=0.4.3,<1.3"
]
test = [
    "coverage>=4.5.3,<7.3",
    "coveralls>=3.0.0,<3.3.2",
    "flake8>=6.0,<6.1",
    "flake8-black>=0.3.4,<0.4",
    "flake8-bugbear>=23.5,<23.8",
    "pytest>=7.1.2,<7.4.2",
    "pytest-subtests>=0.10.0,<0.11.1"
]

[project.scripts]
panoptes_aggregation = "panoptes_aggregation.scripts.aggregation_parser:main"

[project.gui-scripts]
panoptes_aggregation_gui = "panoptes_aggregation.scripts.gui:gui"

[project.urls]
Documentation = "https://aggregation-caesar.zooniverse.org/docs"
Source = "https://github.com/zooniverse/aggregation-for-caesar"

[tool.flit.sdist]
include = [
    "panoptes_aggregation/scripts/icons/*",
    "LICENSE"
]
exclude = [
    "docs/",
    "kubernetes",
    "make_docs.sh"
]

[tool.coverage.run]
omit = [
    "*test*",
    "panoptes_aggregation/scripts/gui*.py",
    "panoptes_aggregation/scripts/no_gooey.py",
    "panoptes_aggregation/scripts/path_type.py"
]
source = ["panoptes_aggregation"]
command_line = "-m pytest"

[tool.coverage.report]
show_missing = true
